+++ packages/vite-plugin/src/client/iife/content-pro-loader.ts
@@ -1,6 +1,9 @@
 declare const __SCRIPT__: string
+const injectTime = performance.now()
 ;(async () => {
-  await import(/* @vite-ignore */ chrome.runtime.getURL(__SCRIPT__))
+  const { onExecute } = await import(/* @vite-ignore */ chrome.runtime.getURL(__SCRIPT__)) as ContentScriptAPI.ModuleExports
+  // this is the entry point of the content script, it will run each time this script is injected
+  onExecute?.({ perf: { injectTime, loadTime: performance.now() - injectTime } })
 })().catch(console.error)
 
 export {}