+++ packages/vite-plugin/tests/e2e/mv3-vite-react-content-script-seq-hmr/src1/main.jsx
@@ -1,0 +1,14 @@
+import React from 'react'
+import ReactDOM from 'react-dom'
+import App from './App'
+
+const root = document.createElement('div')
+root.id = 'root'
+document.body.appendChild(root)
+
+ReactDOM.render(
+  <React.StrictMode>
+    <App />
+  </React.StrictMode>,
+  root,
+)