+++ packages/vite-plugin/tests/e2e/mv3-vite-vanilla-content-script-module-api/src/root.ts
@@ -1,0 +1,16 @@
+const root = document.createElement('div')
+root.id = 'root'
+root.addEventListener('click', () => {
+  chrome.runtime.sendMessage('clicked')
+})
+document.body.append(root)
+
+let injectCount = 0
+export function onExecute({ perf }: ContentScriptAPI.ExecuteFnOptions) {
+  injectCount++
+  root.textContent = `injected ${injectCount}x`
+  root.dataset.injected = injectCount.toString()
+  root.dataset.injectTime = perf.injectTime.toString()
+  root.dataset.loadTime = perf.loadTime.toString()
+  console.log('content script ran')
+}