+++ packages/vite-plugin/tests/e2e/mv3-vite-vanilla-content-script-module-api/src/service-worker.ts
@@ -1,0 +1,10 @@
+import contentScript from './root?script'
+
+chrome.runtime.onMessage.addListener((msg, sender) => {
+  chrome.scripting
+    .executeScript({
+      target: { tabId: sender.tab!.id! },
+      files: [contentScript],
+    })
+    .then(() => console.log('injected'))
+})