+++ packages/vite-plugin/tests/out/content-script-module-api/src/content.js
@@ -1,0 +1,4 @@
+console.log('content script')
+export function onExecute() {
+  console.log('onExecute')
+}