+++ packages/rollup-plugin/CHANGELOG.md
@@ -1,5 +1,19 @@
 # rollup-plugin-chrome-extension
 
+## 3.6.13-beta.0
+
+### Patch Changes
+
+- bbc4d33: fix: schema of input_components
+
+## 3.6.11
+
+## 3.6.11-beta.0
+
+### Patch Changes
+
+- 2fd88aa: Allows <all_urls> match in manifest.json within rollup package
+
 ## 3.6.10
 
 ### Patch Changes